refactor(UI): optimize ctrlSecheduledTest by leveraging Composition and Type Safety
Description:

Leveraged Composition: Removed the local instance of clsLocalDrivingLicenseApplications (_LocalApp). Instead, the control now retrieves applicant data (Full Name, Class Name) directly from the _TestAppointment object using the newly implemented composition properties.

Improved Type Safety: Added a public property TestTypeID using the enTestType enum. Setting this property now automatically triggers UI updates (_SetFormLayout), ensuring the correct image and titles are displayed.

Simplified Logic:

Refactored _SetTestImage and _UpdateGroupBoxTitle to use the internal enum state directly.

Streamlined _FillData to pull information from fewer objects, reducing memory overhead and complexity.

Code Cleanup: Removed redundant database lookups for the local application object within LoadScheduledTestInfo.

+++ DVLD_UI/Test/Controls/ctrlScheduleTest.cs
@@ -202,7 +202,7 @@ namespace DVLD_UI.Test.Controls
             }
 
             _TestTypeFees = clsTestTypes.Find(_TestTypeID).TestTypeFees;
-            _Trial = clsTests.GetTotalTrialsPerTest(_LocalDrivingLicenseApplicationID,(int)_TestTypeID);
+            _Trial = clsTests.GetTotalTrialsPerTest(_LocalDrivingLicenseApplicationID,_TestTypeID);
 
             _FillData();
         }
